cypherToList
cypherToList allows users to return anything from a Cypher query--nodes, relationships, paths, collections, and properties--as it is returned as a list. This is necessary for users who want more than just the cypher function, which returns tabular Cypher results as a data.frame.

cypherToList supplants the following:

* getNodes
* getSingleNode
* getRels
* getSingleRel
* getPaths
* getSinglePath

The above are functions that accept a Cypher query and return a list of the given object. These are no longer necessary, as cypherToList covers all of these cases.

Closes #25.