Allow lists for parameters
For any function that has dots (…) for passing an arbitrary amount of
key=value pairs, make it so this also works if a single list is passed
in place of the dots. That is,

createNode(graph, “Person”, name=“Nicole”, age=24)

can also be written

props = list(name=“Nicole”, age=24)
createNode(graph, “Person”, props)

This has been applied for:

* createNode
* getOrCreateNode
* createRel
* updateProp
* deleteProp
* appendCypher

The tests and docs were also updated, of course. This was not applied
for the following, as they have been deprecated in favor of
cypherToList:

* getNodes
* getSingleNode
* getRels
* getSingleRel
* getPaths
* getSinglePath