Use cypherToList behind the scenes in createNode
This was originally using the /node endpoint to create the node, but
you can’t add labels with this endpoint so it was really clunky to
create the node and then delete it when adding the label failed. This
also gives a better error message when creating a node violates a
constraint.